Many updates

+++ app/src/main/java/com/example/arrow/ProfileActivity.java
@@ -181,7 +181,7 @@ public class ProfileActivity extends AppCompatActivity {
                         String pronoun = String.valueOf(snapshot.child("pronoun").getValue());
                         String lname = String.valueOf(snapshot.child("lName").getValue());
                         String college = String.valueOf(snapshot.child("college").getValue());
-                        int rating = Integer.parseInt(String.valueOf(snapshot.child("overallRating").getValue()));
+                        float rating = Float.parseFloat(String.valueOf(snapshot.child("overallRating").getValue()));
 
                         dataProfs.add(new MyCardHelperClass(pronoun + " " + lname, college, R.drawable.prof_sample, rating));
